+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Deprecated LineItem/UnitTaxAmountMustBeGreaterThanZero error in favor of UnitTaxAmountCannotBeNegative.
 * Add `sourceMerchantId` to `WebhookNotification`s if present
 * Add support for taxAmount field on transaction lineItems
+* Add lineItems method to Transaction
 
 ## 2.5.0
 * Add support for Level 3 summary parameters: `shippingAmount`, `discountAmount`, and `shipsFromPostalCode`